+++ davarocr/davar_table/__init__.py
@@ -4,11 +4,12 @@
 # Filename       :    __init__.py
 # Abstract       :
 
-# Current Version:    1.0.0
-# Date           :    2021-09-18
+# Current Version:    1.0.1
+# Date           :    2022-09-05
 ##################################################################################################
 """
 
 from .models import *
 from .core import *
 from .datasets import *
+from .utils import *