+++ davarocr/davar_table/utils/parallel.py
@@ -1,6 +1,7 @@
 from tqdm import tqdm
 from concurrent.futures import ProcessPoolExecutor, as_completed
 
+
 def parallel_process(array, function, n_jobs=16, use_kwargs=False, front_num=0):
     """
         A parallel version of the map function with a progress bar.
@@ -21,6 +22,7 @@ def parallel_process(array, function, n_jobs=16, use_kwargs=False, front_num=0):
         front = [function(**a) if use_kwargs else function(a) for a in array[:front_num]]
     else:
         front = []
+
     # If we set n_jobs to 1, just run a list comprehension. This is useful for benchmarking and debugging.
     if n_jobs == 1:
         return front + [function(**a) if use_kwargs else function(a) for a in tqdm(array[front_num:])]